refactor(CountdownTextView): move tests to Robolectric
