+++ app/build.gradle
@@ -58,6 +58,12 @@ android {
   kotlinOptions {
     jvmTarget = 1.8
   }
+
+  testOptions {
+    unitTests {
+      includeAndroidResources = true
+    }
+  }
 }
 
 dependencies {
@@ -74,6 +80,10 @@ dependencies {
   implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 
   playstoreImplementation 'com.google.android.gms:play-services-cast-framework:18.1.0'
+
+  testImplementation 'io.mockk:mockk:1.10.0'
+  testImplementation 'junit:junit:4.13'
+  testImplementation 'org.robolectric:robolectric:4.3.1'
 
   androidTestImplementation 'androidx.test:core:1.2.0'
   androidTestImplementation 'androidx.test:rules:1.2.0'
